refactor: rest protocol
