+++ protocol/rest/server/go_restful_server.go
@@ -15,11 +15,10 @@
  * limitations under the License.
  */
 
-package server_impl
+package server
 
 import (
 	"context"
-
 	"dubbo.apache.org/dubbo-go/v3/config"
 	"fmt"
 	"net"
@@ -36,15 +35,12 @@ import (
 
 import (
 	"dubbo.apache.org/dubbo-go/v3/common"
-	"dubbo.apache.org/dubbo-go/v3/common/constant"
-	"dubbo.apache.org/dubbo-go/v3/common/extension"
 	"dubbo.apache.org/dubbo-go/v3/common/logger"
-	"dubbo.apache.org/dubbo-go/v3/protocol/rest/server"
 )
 
-func init() {
-	extension.SetRestServer(constant.DefaultRestServer, NewGoRestfulServer)
-}
+//func init() {
+//	rest.SetRestServer(constant.DefaultRestServer, NewGoRestfulServer)
+//}
 
 var filterSlice []restful.FilterFunction
 
@@ -55,7 +51,7 @@ type GoRestfulServer struct {
 }
 
 // NewGoRestfulServer a constructor of GoRestfulServer
-func NewGoRestfulServer() server.RestServer {
+func NewGoRestfulServer() RestServer {
 	return &GoRestfulServer{}
 }
 
@@ -88,7 +84,7 @@ func (grs *GoRestfulServer) Start(url *common.URL) {
 
 // Publish a http api in go-restful server
 // The routeFunc should be invoked when the server receive a request
-func (grs *GoRestfulServer) Deploy(restMethodConfig *config.RestMethodConfig, routeFunc func(request server.RestServerRequest, response server.RestServerResponse)) {
+func (grs *GoRestfulServer) Deploy(restMethodConfig *config.RestMethodConfig, routeFunc func(request RestServerRequest, response RestServerResponse)) {
 	rf := func(req *restful.Request, resp *restful.Response) {
 		routeFunc(NewGoRestfulRequestAdapter(req), resp)
 	}
@@ -124,7 +120,7 @@ func AddGoRestfulServerFilter(filterFuc restful.FilterFunction) {
 
 // GoRestfulRequestAdapter a adapter struct about RestServerRequest
 type GoRestfulRequestAdapter struct {
-	server.RestServerRequest
+	RestServerRequest
 	request *restful.Request
 }
